# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RealtimeAlgoChatApp Studio API — secure rooms, profiles, and E2EE collaboration
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @maxLength 16
 */
export type UserProfileAvatar = typeof UserProfileAvatar[keyof typeof UserProfileAvatar] | null;


export const UserProfileAvatar = {
  '🧑‍💻': '🧑‍💻',
  '🚀': '🚀',
  '🦊': '🦊',
  '🐼': '🐼',
  '🐸': '🐸',
  '🐙': '🐙',
  '🦄': '🦄',
  '🌻': '🌻',
  '🍕': '🍕',
  '🎮': '🎮',
  '🔥': '🔥',
  '💎': '💎',
} as const;

export interface UserProfile {
  userId: string;
  username: string;
  /** @maxLength 16 */
  avatar?: UserProfileAvatar;
  /** Base64-encoded X25519 public key for E2EE */
  publicKey?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileResponse {
  profile: UserProfile;
}

/**
 * @maxLength 16
 */
export type UpsertProfileRequestAvatar = typeof UpsertProfileRequestAvatar[keyof typeof UpsertProfileRequestAvatar] | null;


export const UpsertProfileRequestAvatar = {
  '🧑‍💻': '🧑‍💻',
  '🚀': '🚀',
  '🦊': '🦊',
  '🐼': '🐼',
  '🐸': '🐸',
  '🐙': '🐙',
  '🦄': '🦄',
  '🌻': '🌻',
  '🍕': '🍕',
  '🎮': '🎮',
  '🔥': '🔥',
  '💎': '💎',
} as const;

export interface UpsertProfileRequest {
  /** @maxLength 64 */
  username?: string;
  /** @maxLength 16 */
  avatar?: UpsertProfileRequestAvatar;
  /**
     * Base64-encoded X25519 public key
     * @maxLength 64
     */
  publicKey?: string | null;
}

export interface Room {
  id: string;
  name: string;
  userCount?: number;
  /** Unix timestamp in milliseconds */
  createdAt: number;
}

export interface RoomsListResponse {
  rooms: Room[];
}

export interface CreateRoomRequest {
  /** @maxLength 128 */
  roomId: string;
  /** @maxLength 128 */
  roomName?: string;
}

export interface RoomResponse {
  room: Room;
}

export type UnauthorizedResponse = {
  error: string;
};