# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RealtimeAlgoChatApp Studio API — secure rooms, profiles, and E2EE collaboration
 * OpenAPI spec version: 0.2.0
 */
import type { UserProfile } from './userProfile';

export interface ProfileResponse {
  profile: UserProfile;
}